Add parsing of TTL "numbers"
While there is no set standard for what is/is not allowed in a zonefile,
BIND allows you to parse the following:

 #s = seconds = # x 1 seconds
 #m = minutes = # x 60 seconds
 #h = hours = # x 3600 seconds
 #d = day = # x 86400 seconds
 #w = week = # x 604800 seconds

This also allows for things like "10h5m" to specify that it is 10 hours
and 5 minutes.

The parser converts this to an i32 internally before returning it as a
token.